HDLC lenght can be 11 bits.
GetDifference method added for GXDateTime.
HDLC multiple client frames issue fixed.

+++ Development/Internal/GXCommon.cs
@@ -50,8 +50,8 @@ namespace Gurux.DLMS.Internal
         internal const byte HDLCFrameStartEnd = 0x7E;
         internal static readonly byte[] LogicalNameObjectID = { 0x60, 0x85, 0x74, 0x05, 0x08, 0x01, 0x01 };
         internal static readonly byte[] ShortNameObjectID = { 0x60, 0x85, 0x74, 0x05, 0x08, 0x01, 0x02 };
-        internal static readonly byte[] LogicalNameObjectIdWithCiphering = { 0x60, (byte)0x85, 0x74, 0x05, 0x08, 0x01, 0x03 };
-        internal static readonly byte[] ShortNameObjectIdWithCiphering = { 0x60, (byte)0x85, 0x74, 0x05, 0x08, 0x01, 0x04 };
+        internal static readonly byte[] LogicalNameObjectIdWithCiphering = { 0x60, 0x85, 0x74, 0x05, 0x08, 0x01, 0x03 };
+        internal static readonly byte[] ShortNameObjectIdWithCiphering = { 0x60, 0x85, 0x74, 0x05, 0x08, 0x01, 0x04 };
 
         /// <summary>
         /// Sent LLC bytes.
